requirements.txt
Flask==2.3.3
Flask-APScheduler==1.12.4
Werkzeug==2.3.7
PyMySQL==1.1.0
python-dateutil==2.8.2
tzlocal==5.0.1
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.1.1
redis==4.6.0